requirements.txt
flask
google_auth_oauthlib
google-api-python-client
google-cloud-ndb
markupsafe
jinja2
